Added client updater & fixed small bugs

+++ client/debug.cpp
@@ -16,10 +16,23 @@ boost::shared_ptr<addonDebug> gDebug;
 
 addonDebug::addonDebug()
 {
+	boost::filesystem::path log(".\\addon.log");
+
+	if(boost::filesystem::exists(log))
+	{
+		boost::system::error_code error;
+
+		boost::filesystem::remove(log, error);
+
+		//if(error)
+			//gDebug->Log("Cannot remove old log file: %s (Error code: %i)", error.message().c_str(), error.value());
+	}
+
 	this->mutexInstance = boost::shared_ptr<boost::mutex>(new boost::mutex());
 	this->threadInstance = boost::shared_ptr<boost::thread>(new boost::thread(boost::bind(&addonDebug::Thread)));
 
 	this->traceLastFunction("addonDebug::addonDebug() at 0x?????");
+	this->Log("Called debug constructor");
 }
 
 
@@ -27,31 +40,39 @@ addonDebug::addonDebug()
 addonDebug::~addonDebug()
 {
 	this->traceLastFunction("addonDebug::~addonDebug() at 0x?????");
+	this->Log("Called debug destructor");
 
-	this->getMutexInstance()->destroy();
-	this->getThreadInstance()->interrupt();
+	this->threadInstance->interruption_requested();
+	this->mutexInstance->destroy();
 }
 
 
 
 LONG WINAPI addonDebug::UnhandledExceptionFilter(struct _EXCEPTION_POINTERS *ExceptionInfo)
 {
-	boost::system::error_code error;
+	static bool processingException = false;
+
+	if(processingException)
+		return EXCEPTION_CONTINUE_EXECUTION;
+
+	processingException = true;
 
 	boost::filesystem::path log(".\\addon_crashreport.log");
 
 	if(boost::filesystem::exists(log))
 	{
+		boost::system::error_code error;
+
 		boost::filesystem::remove(log, error);
 
 		if(error)
-			gDebug->Log("Cannot remove old crashreport file: %s (%i)", error.message().c_str(), boost::lexical_cast<int>(error));
+			gDebug->Log("Cannot remove old crashreport file: %s (Error code: %i)", error.message().c_str(), error.value());
 	}
 
 	std::fstream report;
 
 	report.open("addon_crashreport.log", (std::fstream::out | std::fstream::app));
-	//report.setf(std::ios_base::hex | std::ios_base::uppercase);
+
 	report << "-----------------------------------------------------------------" << std::endl;
 	report << "\t\t     SAMP-Addon was crashed" << std::endl;
 	report << "-----------------------------------------------------------------" << std::endl << std::endl;
@@ -205,19 +226,12 @@ LONG WINAPI addonDebug::UnhandledExceptionFilter(struct _EXCEPTION_POINTERS *Exc
 	report << std::endl;
 	report << "Backtrace:" << std::endl << std::endl;
 
-	int c = 0;
-
-	std::reverse(gDebug->funcTrace.begin(), gDebug->funcTrace.end());
-
-	for(std::vector<std::string>::iterator i = gDebug->funcTrace.begin(); i != gDebug->funcTrace.end(); i++)
+	for(std::list<std::string>::iterator i = gDebug->funcTrace.begin(); i != gDebug->funcTrace.end(); i++)
 	{
-		if(!c)
+		if(i == gDebug->funcTrace.begin())
 			report << (*i) << std::endl;
 		else
 			report << "<= " << (*i) << std::endl;
-
-		if(++c > 20)
-			break;
 	}
 
 	report << std::endl << std::endl;
@@ -226,7 +240,7 @@ LONG WINAPI addonDebug::UnhandledExceptionFilter(struct _EXCEPTION_POINTERS *Exc
 
 	gDebug->Log("Exception 0x%08x at address 0x%08x. Addon was terminated. See 'addon_crashreport.log'", ExceptionInfo->ExceptionRecord->ExceptionCode, ExceptionInfo->ExceptionRecord->ExceptionAddress);
 
-	boost::this_thread::sleep_for(boost::chrono::seconds(1));
+	boost::this_thread::sleep_for(boost::chrono::milliseconds(500));
 	exit(EXIT_FAILURE);
 
 	return EXCEPTION_CONTINUE_SEARCH;
@@ -236,17 +250,18 @@ LONG WINAPI addonDebug::UnhandledExceptionFilter(struct _EXCEPTION_POINTERS *Exc
 
 void addonDebug::Log(char *format, ...)
 {
-	//this->traceLastFunction("addonDebug::Log(...) at 0x%x", &addonDebug::Log);
-
 	va_list args;
+	boost::mutex logMutex;
 
 	va_start(args, format);
 
-	this->getMutexInstance()->lock();
-	this->Queue.push(addonString::vprintf(format, args));
-	this->getMutexInstance()->unlock();
+	logMutex.lock();
+	this->logQueue.push(addonString::vprintf(format, args));
+	logMutex.unlock();
 
 	va_end(args);
+
+	logMutex.destroy();
 }
 
 
@@ -254,14 +269,24 @@ void addonDebug::Log(char *format, ...)
 void addonDebug::traceLastFunction(char *format, ...)
 {
 	va_list args;
+	boost::mutex traceMutex;
 
 	va_start(args, format);
 
-	this->getMutexInstance()->lock();
-	this->funcTrace.push_back(addonString::vprintf(format, args));
-	this->getMutexInstance()->unlock();
+	traceMutex.lock();
+	this->funcTrace.push_front(addonString::vprintf(format, args));
+	traceMutex.unlock();
 
 	va_end(args);
+
+	if(this->funcTrace.size() > 20)
+	{
+		traceMutex.lock();
+		this->funcTrace.pop_back();
+		traceMutex.unlock();
+	}
+
+	traceMutex.destroy();
 }
 
 
@@ -271,6 +296,7 @@ void addonDebug::Thread()
 	assert(gDebug->getThreadInstance->get_id() == boost::this_thread::get_id());
 
 	gDebug->traceLastFunction("addonDebug::Thread() at 0x%x", &addonDebug::Thread);
+	gDebug->Log("Started file debug thread with id 0x%x", gDebug->getThreadInstance()->get_thread_info()->id);
 
 	char timeform[16];
 	struct tm *timeinfo;
@@ -278,18 +304,17 @@ void addonDebug::Thread()
 
 	std::fstream file;
 	std::string data;
-	boost::mutex localMutex;
 
 	while(true)
 	{
-		while(!gDebug->Queue.empty())
+		while(!gDebug->logQueue.empty())
 		{
 			boost::this_thread::disable_interruption di;
 
-			localMutex.lock();
-			data = gDebug->Queue.front();
-			gDebug->Queue.pop();
-			localMutex.unlock();
+			gDebug->getMutexInstance()->lock();
+			data = gDebug->logQueue.front();
+			gDebug->logQueue.pop();
+			gDebug->getMutexInstance()->unlock();
 
 			time(&rawtime);
 			timeinfo = localtime(&rawtime);
@@ -299,6 +324,7 @@ void addonDebug::Thread()
 			file << "[" << timeform << "] " << data << std::endl;
 			file.close();
 
+			boost::this_thread::restore_interruption re(di);
 			boost::this_thread::sleep_for(boost::chrono::milliseconds(1));
 		}
